core: Extract cell structure communication
