+++ src/core/communication.cpp
@@ -31,9 +31,7 @@
 
 #include "errorhandling.hpp"
 
-#include "CellStructure.hpp"
 #include "EspressoSystemInterface.hpp"
-#include "cells.hpp"
 #include "cuda_interface.hpp"
 #include "energy.hpp"
 #include "event.hpp"
@@ -316,18 +314,6 @@ void mpi_bcast_coulomb_params_slave(int, int) {
 #if defined(ELECTROSTATICS) || defined(DIPOLES)
   on_coulomb_change();
 #endif
-}
-
-/*************** BCAST CELL STRUCTURE *****************/
-
-REGISTER_CALLBACK(cells_re_init)
-
-void mpi_bcast_cell_structure(int cs) { mpi_call_all(cells_re_init, cs); }
-
-REGISTER_CALLBACK(cells_set_use_verlet_lists)
-
-void mpi_set_use_verlet_lists(bool use_verlet_lists) {
-  mpi_call_all(cells_set_use_verlet_lists, use_verlet_lists);
 }
 
 /******************* BCAST CUDA GLOBAL PART VARS ********************/
